+++ src/rephemeral/__init__.py
@@ -8,4 +8,4 @@ Free Software Foundation, either version 3 of the License, or (at your
 option) any later version. See the LICENSE file for the full text.
 """
 
-__version__ = "0.5.0"
+__version__ = "0.6.0"